+++ Dockerfile
@@ -16,3 +16,7 @@ COPY . /app/
 WORKDIR /app
 EXPOSE 8000/tcp
 CMD ["poetry", "run", "gunicorn", "-b", "0.0.0.0", "app:create_app()"]
+
+FROM base as test
+WORKDIR /app
+CMD ["poetry", "run", "pytest"]
