+++ poetry.toml
@@ -1,3 +1,3 @@
 [virtualenvs]
 create = true
-in-project = false
+in-project = true
